pyproject.toml
[tool.poetry]
name = "image-validator"
version = "0.1.0"
description = ""
authors = ["Max Herbert <max.herbert@diamond.ac.uk>"]
readme = "README.md"


[tool.poetry.scripts]
image-validator = "image_validator.main:app"

[tool.poetry.dependencies]
python = "^3.10"
typer = "^0.12.5"
pyyaml = "^6.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
